Update README.md
Banks have to refill the ATMs when the money goes below a specific threshold limit. This depends on the activity and the area where a particular ATM is located as well as the weather, day of the week, etc. 

The data in this project is from more than 100 ATMs across Denmark. Data is captured for every transaction including, card type, location, date, time, ATM type, etc.

Spar Nord Bank was trying to observe the withdrawal behavior and the corresponding dependent factors to optimally manage the refill frequency. Apart from this, some business insights had to be drawn from the data.

This project built a batch ETL pipeline to extract (Sqoop, HDFS) transactional data from RDS, transform (PySpark), and load (S3, Redshift) it into target dimensions and fact on a Redshift Data Mart.